template do tempo

+++ include/search.h
@@ -7,7 +7,7 @@
 #include <cmath>
 #include <algorithm>
 
-int *lin_search(int *first, int *last, int target);
+int *lin_search(int *first, int *last, short target);
 int *bsearch(int *first, int *last, int target);
 
 #endif
